Sync updates from SVN: numerical methods overhaul, bug fixes, new unit tests
- numerical: add value_fun_g/rvalue_fun interfaces, D1/D2 derivative helpers,
  improved Newton-Raphson convergence criterion (scale-independent step test)
- numerical_sm: add linked_list_sm and numerical_sm as tracked submodule files
- rpn_stack_sm: fix to_cartesian(b) -> to_cartesian(a) in arithmetic ops,
  fix sqrt/cosine/atan2 precision issues
- hp/main: minor fixes (error message, nint rounding, mem_block_size constant)
- tests: add utest_amap, utest_linked_list, utest_numerical, test_amap

+++ src/amap.f90
@@ -1,6 +1,5 @@
-! Associative map string -> real(real64)
+! Associative map string -> real(8)
 module amap
-    use iso_fortran_env, only: real64
     implicit none
     
     ! The key
@@ -19,7 +18,7 @@ module amap
     ! The value
     type value_t
         private
-        real(real64) :: v = huge(0.0d0) ! An out-of-band value
+        real(8) :: v = huge(0.0d0) ! An out-of-band value
     contains
         procedure, private :: write_value_t
         generic, public    :: write(formatted) => write_value_t
@@ -78,7 +77,7 @@ contains
     
     subroutine set_to_value_t(this, v)
         class(value_t), intent(inout)  :: this
-        real(real64), intent(in) :: v
+        real(8), intent(in) :: v
         this%v = v
     end subroutine set_to_value_t
 
@@ -94,7 +93,7 @@ contains
     subroutine set_amap_t(this,kv,vv)
         class(amap_t), intent(inout) :: this
         character(len=*), intent(in) :: kv
-        real(real64), intent(in)          :: vv
+        real(8), intent(in)          :: vv
         type(pair_t), allocatable :: tmp_pairs(:)
         type(key_t)               :: k
         type(value_t)             :: v
@@ -152,7 +151,7 @@ contains
     function get_value_amap_t(this, kv) result(r)
         class(amap_t), intent(in)    :: this
         character(len=*), intent(in) :: kv
-        real(real64) :: r
+        real(8) :: r
         type(value_t) :: s
         s = this%get(kv)
         r = s%v
@@ -181,7 +180,7 @@ contains
     
     function equals_value_t(this, v) result(r)
         class(value_t), intent(in) :: this
-        real(real64), intent(in) :: v
+        real(8), intent(in) :: v
         logical :: r
         r = this%v == v
     end function equals_value_t
